Use the `client_token` option with the openmetrics v2 implementation (#12764)
* Use the `client_token` option with the openmetrics v2 implementation

* Use `values()` instead of `item()`

* Move the e2e tests to their own file
