+++ vault/tests/conftest.py
@@ -41,9 +41,15 @@ def global_tags():
 
 @pytest.fixture(scope='session')
 def instance(dd_get_state):
-    def get_instance():
+    def get_instance(use_auth_file=True):
         inst = INSTANCES['main'].copy()
-        inst['client_token_path'] = dd_get_state('client_token_path')
+
+        if use_auth_file:
+            inst['client_token_path'] = dd_get_state('client_token_path')
+        else:
+            with open(dd_get_state('client_token_path'), 'r') as auth_file:
+                inst['client_token'] = auth_file.read()
+
         return inst
 
     return get_instance
@@ -57,10 +63,18 @@ def no_token_instance():
 
 
 @pytest.fixture(scope='session')
-def e2e_instance():
-    inst = INSTANCES['main'].copy()
-    inst['client_token_path'] = '/home/vault-sink/token'
-    return inst
+def e2e_instance(dd_get_state):
+    def get_instance(use_auth_file=True):
+        inst = INSTANCES['main'].copy()
+
+        if use_auth_file:
+            inst['client_token_path'] = '/home/vault-sink/token'
+        else:
+            with open(dd_get_state('client_token_path'), 'r') as auth_file:
+                inst['client_token'] = auth_file.read()
+        return inst
+
+    return get_instance
 
 
 @pytest.fixture(scope='session')
@@ -87,7 +101,7 @@ def dd_environment(e2e_instance, dd_save_state):
         ):
             dd_save_state('client_token_path', token_file)
 
-            yield e2e_instance, {'docker_volumes': ['{}:/home/vault-sink'.format(sink_dir)]}
+            yield e2e_instance(), {'docker_volumes': ['{}:/home/vault-sink'.format(sink_dir)]}
 
 
 class ApplyPermissions(LazyFunction):
